movies_admin/requirements/base.txt
django==3.2.6
django-model-utils==4.0.0  # Набор полезных базовых классов и утилит для Django
djangorestframework==3.12.4
django-filter==2.4.0
psycopg2-binary==2.8.6
python-dotenv==0.17.1
loguru~=0.5.3
tqdm~=4.62.0
